Clean up the async/await threading
This was causing warnings about the bindings b/c of the `goForward()` method being
called on a background thread in the completion after speaking the last instruction.

Also, this change allows for both the tremor and tapping assessments to use the same
code for tracking the remaining time as well as starting and stopping the recorder.